added gruvbox, catppuccin, rose-pine. Fixed leap.
Leap was bound to 'l', very annoying when trying to move right. Changed
binding to 'm' for 'move to'. Configured gruvbox, catpuccin, and
rose-pine to automatically change variants based on the vim.o.background
setting in vim-options.lua. Added a function to change vim.o.background
based on the time of day in vim-options.lua.